refactor(comments): extract comment pipeline helpers
